requirements.txt
# Core Testing
pytest==8.3.4
pytest-playwright==0.6.2
playwright==1.49.1
pytest-asyncio==0.25.2
pytest-xdist==3.6.1          # Parallel test execution
pytest-html==4.1.1            # HTML test reports
pytest-rerunfailures==14.0    # Retry failed tests

# API Testing
requests==2.32.3
httpx==0.28.1                 # Async HTTP client

# Azure DevOps Integration
azure-devops==7.1.0b4
msrest==0.7.1

# Confluence / Atlassian Integration
atlassian-python-api==3.41.16

# Reporting
allure-pytest==2.13.5
jinja2==3.1.5                 # HTML report templating

# Utilities
python-dotenv==1.0.1          # .env file support
pydantic==2.10.4              # Data validation & settings
pyyaml==6.0.2                 # YAML config support
tenacity==9.0.0               # Retry logic
loguru==0.7.3                 # Structured logging
faker==33.1.0                 # Fake test data generation

# Visual Regression
Pillow==11.1.0               # Screenshot pixel-diff comparison

# Type checking (dev)
mypy==1.14.1
